Рефакторинг на уровне классов
